Optimize the 'join' command to 'team'
